Scratchnotes
This is an initial test for this  not quite building yet but seems like it may. The seperate bridge file still needs to be built, this may end up as a container to host the vst file. Also contains VST2.x sdk files - I am guessing this is not licencenable, just currently testing this, the vst2.x files are available elsewhere on github, not totally familiar with the situation on them I may need to remove them but still need to read up on that. However I am thinking the actual vst2 or vst1 sdk files will be needed - or rebuilt which is a whole different issue.